+++ src/lob_asset_management/models/file.clj
@@ -1,5 +1,5 @@
 (ns lob-asset-management.models.file
   (:require [schema.core :as s]))
 
-(def list-file-name #{:portfolio :transaction :asset :read-release})
+(def list-file-name #{:portfolio :transaction :asset :read-release :telegram})
 (s/defschema file-name (apply s/enum list-file-name))